factorized game
