+++ src/games/tic-tac-toe/components/square/square.js
@@ -1,0 +1,7 @@
+export default function Square({ value, onSquareClick }) {
+    return (
+      <button className="square" onClick={onSquareClick}>
+        {value}
+      </button>
+    );
+  }
